Simplify encoding detection code
The lxml parser can detect the encoding from the header by itself and artist and title can be read from parsed comments. We don't need the manual encoding detection code anymore. If the parser is given a None encoding, it will first try to determine the encoding from the header and if it fails, it will default to UTF-8.