Convert entire SingStar disc at once
In addition to melody.xml, the parser now also parses config.xml for edition, covers.xml for cover file info and songs.xml for song metadata and IDs. It can handle missing files and elements when they are not necessary. It creates song directories for all songs and symlinks a video and a cover file to it. The input to converter.py is now a path to the `filesystem/export` directory found in extracted SingStar discs.